+++ src/main/java/main/view/Project/APITokenServlet.java
@@ -1,18 +1,16 @@
 package main.view.Project;
 
 import main.Session;
-import main.model.dto.DtoMapper;
-import main.model.dto.ImportTokenDto;
+import main.model.dto.APITokenDto;
 import main.view.BaseServlet;
 import main.view.IGet;
-import main.view.IPost;
 
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/project/importToken")
-public class ImportTokenServlet extends BaseServlet implements IGet {
+@WebServlet("/project/apiToken")
+public class APITokenServlet extends BaseServlet implements IGet {
 
     @Override
     public void doGet(HttpServletRequest req, HttpServletResponse resp) {
@@ -21,9 +19,9 @@ public class ImportTokenServlet extends BaseServlet implements IGet {
 
         try {
             Session session = createSession(req);
-            ImportTokenDto tokenDto = new ImportTokenDto();
+            APITokenDto tokenDto = new APITokenDto();
             tokenDto.getSearchTemplateFromRequestParameters(req);
-            tokenDto.setImport_token(session.controllerFactory.getHandler(tokenDto).create(tokenDto).getImport_token());
+            tokenDto.setApi_token(session.controllerFactory.getHandler(tokenDto).create(tokenDto).getApi_token());
             resp.getWriter().write(mapper.serialize(tokenDto));
         } catch (Exception e) {
             handleException(resp, e);